Currency Update for Home Screen Theme update for light, dark

+++ app/src/main/java/com/dsk/myexpense/expense_module/data/repository/ExpenseRepository.kt
@@ -31,9 +31,9 @@ class ExpenseRepository(
 ) {
 
     val allExpenseDetails: LiveData<List<ExpenseDetails>> = expenseDAO.getAllExpenseDetails()
-    val getTotalIncomeAmount: LiveData<Int> = expenseDAO.getTotalIncome().asLiveData()
-    val getTotalExpenseAmount: LiveData<Int> = expenseDAO.getTotalExpense().asLiveData()
-    val getTotalIncomeExpenseAmount: LiveData<Int> = expenseDAO.getTotalIncomeExpense().asLiveData()
+    val getTotalIncomeAmount: LiveData<Double> = expenseDAO.getTotalIncome().asLiveData()
+    val getTotalExpenseAmount: LiveData<Double> = expenseDAO.getTotalExpense().asLiveData()
+    val getTotalIncomeExpenseAmount: LiveData<Double> = expenseDAO.getTotalIncomeExpense().asLiveData()
 
     suspend fun saveExpenseWithInvoice(
         expenseDetails: ExpenseDetails, categoryName: String, bitmap: Bitmap?
